+++ examples/android/src/platform/graphics.h
@@ -1,0 +1,14 @@
+#include "../base.h"
+
+namespace G {
+
+int CreateProgram() {
+    int ph = glCreateProgram();
+    if (ph == 0) {
+        LOGE("glCreateProgram RETURNED 0! PROGRAM CREATION FAILED.");
+    }
+
+    return ph;
+}
+
+}